packages/br_sdk/pyproject.toml
[tool.uv]
package = true

[project]
name = "br_sdk"
version = "0.1.0"
dependencies = [
    "blinker>=1.9.0",
    "pydantic>=2.11.4",
    "pyyaml>=6.0.2",
]

[tool.pytest.ini_options]
pythonpath = [
  "src"
]
